upload keystrokeUI.py
Display real-time keyboard key actuation events. Basic visualization application without the risks of running executables from the web. Using tkinter and functional programming style it is very intuitive to adjust and modify. The visualizer window can be dragged and moved around anywhere on the screen and a close button is available with window sizes similar to default. Run "python keystrokeUI.py" in any console or terminal environment to start the application.