Add files via upload
Run this in your memes and other downloaded pictures folder to convert all .webp files in the current folder to a normal .png file. It is fully portable for any OS that has a relatively modern version of python installed. You will likely need to run "pip install PIL", as this module is the core of the script.