Add files via upload
Visualiser and a calculator for the data-frame transfer process for the HDLC (LAPB) communication protocol. It requires a few basic inputs related to total data package size, window size limit, single frame size limit, time delay and custom selectable failed frame. The visualization is driven by the calculations that take into account all the specifics and edge-cases of the exam to which this script came as a solution. This is an amazing tool for learning and education, as it efficiently conveys the meaning of a "sliding-window" communication system and Go-Back-N error correction systems and can easily be related to the specifics of the hardware implementation of the communication system. This script has been written and verified on Matlab version r2013a.